removed require jquery from jquery.ui.core
I get why it's there in the first place, but there are a *lot* of
use cases where you wouldn't want your jquery library included
inline with everything else (preferring to link to a common cdn
version ala google or cdnjs, etc.). Anyhow, it strikes me as the
sort of thing where you're better off giving people the opportunity
to choose whether or not to require it themselves.